My solution3

+++ app/main.py
@@ -5,7 +5,7 @@ class Car:
     def __init__(
         self,
         comfort_class: int,
-        clean_mark: float,
+        clean_mark: int,
         brand: str,
     ) -> None:
         self.comfort_class = comfort_class
